backend/requirements.txt
annotated-doc
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
bcrypt
certifi
cffi
click
coverage
dnspython
ecdsa
email-validator
fastapi
h11
httpcore
httpx
idna
iniconfig
packaging
passlib
pluggy
pyasn1
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
pytest
pytest-asyncio
pytest-cov
python-dotenv
python-jose
python-multipart
rsa
six
SQLAlchemy
starlette
typing-inspection
typing_extensions
uvicorn
psycopg2-binary
sqlalchemy